backend/requirements.txt
fastapi
uvicorn
catboost
pandas
python-dotenv
google-generativeai
